Tweaking comments and clarity
