+++ pkg/shared/ivcs.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/go-plugin"
 	"github.com/scan-io-git/scan-io/pkg/shared/config"
 
-	ftutils "github.com/scan-io-git/scan-io/internal/fetcher-utils"
+	ftutils "github.com/scan-io-git/scan-io/internal/fetcherutils"
 )
 
 // RepositoryParams holds the details of a repository.
